Refactor ctd detector
